refactor(csrc): split args and difftest from npc_main
