+++ npc/csrc/npc_main.cpp
@@ -1,4 +1,6 @@
 #include "Vtb_triathlon.h"
+#include "args_parser.h"
+#include "difftest_client.h"
 #include "verilated.h"
 #include "verilated_vcd_c.h"
 
@@ -8,7 +10,6 @@
 #include <cstdlib>
 #include <cstring>
 #include <deque>
-#include <dlfcn.h>
 #include <fstream>
 #include <iostream>
 #include <string>
@@ -26,342 +27,6 @@ constexpr uint32_t kPmemSize = 0x08000000u;
 constexpr uint32_t kMmioBase = 0xA0000000u;
 constexpr uint32_t kMmioEnd = 0xAFFFFFFFu;
 constexpr uint32_t kSeed4Addr = 0x80003C3Cu;
-
-struct SimArgs {
-  std::string img_path;
-  uint64_t max_cycles = 600000000;
-  std::string difftest_so;
-  bool trace = false;
-  std::string trace_path = "npc.vcd";
-  bool commit_trace = false;
-  bool fe_trace = false;
-  bool bru_trace = false;
-  bool stall_trace = false;
-  uint64_t stall_threshold = 200;
-  uint64_t progress_interval = 0;
-};
-
-static bool parse_u64(const std::string &s, uint64_t &out) {
-  try {
-    size_t idx = 0;
-    out = std::stoull(s, &idx, 0);
-    return idx == s.size();
-  } catch (...) {
-    return false;
-  }
-}
-
-static SimArgs parse_args(int argc, char **argv) {
-  SimArgs args;
-  for (int i = 1; i < argc; i++) {
-    std::string arg = argv[i];
-
-    if (arg == "-d") {
-      if (i + 1 < argc) {
-        args.difftest_so = argv[i + 1];
-        i++;
-      }
-      continue;
-    }
-    if (arg.rfind("--difftest=", 0) == 0) {
-      args.difftest_so = arg.substr(std::string("--difftest=").size());
-      continue;
-    }
-    if (arg == "--max-cycles" && i + 1 < argc) {
-      uint64_t v = 0;
-      if (parse_u64(argv[i + 1], v)) {
-        args.max_cycles = v;
-        i++;
-        continue;
-      }
-    }
-    if (arg.rfind("--max-cycles=", 0) == 0) {
-      uint64_t v = 0;
-      if (parse_u64(arg.substr(std::string("--max-cycles=").size()), v)) {
-        args.max_cycles = v;
-      }
-      continue;
-    }
-    if (arg == "--trace") {
-      args.trace = true;
-      if (i + 1 < argc && argv[i + 1][0] != '-') {
-        args.trace_path = argv[i + 1];
-        i++;
-      }
-      continue;
-    }
-    if (arg.rfind("--trace=", 0) == 0) {
-      args.trace = true;
-      args.trace_path = arg.substr(std::string("--trace=").size());
-      continue;
-    }
-    if (arg == "--commit-trace") {
-      args.commit_trace = true;
-      continue;
-    }
-    if (arg == "--fe-trace") {
-      args.fe_trace = true;
-      continue;
-    }
-    if (arg == "--bru-trace") {
-      args.bru_trace = true;
-      continue;
-    }
-    if (arg == "--stall-trace") {
-      args.stall_trace = true;
-      if (i + 1 < argc) {
-        uint64_t v = 0;
-        if (parse_u64(argv[i + 1], v)) {
-          args.stall_threshold = v;
-          i++;
-        }
-      }
-      continue;
-    }
-    if (arg.rfind("--stall-trace=", 0) == 0) {
-      args.stall_trace = true;
-      uint64_t v = 0;
-      if (parse_u64(arg.substr(std::string("--stall-trace=").size()), v)) {
-        args.stall_threshold = v;
-      }
-      continue;
-    }
-    if (arg == "--progress") {
-      args.progress_interval = 1000000;
-      if (i + 1 < argc && argv[i + 1][0] != '-') {
-        uint64_t v = 0;
-        if (parse_u64(argv[i + 1], v)) {
-          args.progress_interval = v;
-          i++;
-        }
-      }
-      continue;
-    }
-    if (arg.rfind("--progress=", 0) == 0) {
-      uint64_t v = 0;
-      if (parse_u64(arg.substr(std::string("--progress=").size()), v)) {
-        args.progress_interval = v;
-      }
-      continue;
-    }
-    if (!arg.empty() && arg[0] == '-') {
-      continue;
-    }
-    args.img_path = arg;
-  }
-  return args;
-}
-
-struct DifftestCPUState {
-  uint32_t gpr[16];
-  uint32_t pc;
-  struct {
-    uint32_t mtvec;
-    uint32_t mepc;
-    uint32_t mstatus;
-    uint32_t mcause;
-  } csr;
-};
-
-struct DUTCSRState {
-  uint32_t mtvec;
-  uint32_t mepc;
-  uint32_t mstatus;
-  uint32_t mcause;
-};
-
-class Difftest {
- public:
-  bool init(const std::string &so_path,
-            const std::vector<uint32_t> &pmem_words,
-            uint32_t entry_pc) {
-    handle_ = dlopen(so_path.c_str(), RTLD_LAZY);
-    if (!handle_) {
-      std::cerr << "[difftest] dlopen failed: " << dlerror() << "\n";
-      return false;
-    }
-
-    difftest_memcpy_ =
-        reinterpret_cast<difftest_memcpy_t>(dlsym(handle_, "difftest_memcpy"));
-    difftest_regcpy_ =
-        reinterpret_cast<difftest_regcpy_t>(dlsym(handle_, "difftest_regcpy"));
-    difftest_exec_ =
-        reinterpret_cast<difftest_exec_t>(dlsym(handle_, "difftest_exec"));
-    difftest_init_ =
-        reinterpret_cast<difftest_init_t>(dlsym(handle_, "difftest_init"));
-
-    if (!difftest_memcpy_ || !difftest_regcpy_ || !difftest_exec_ ||
-        !difftest_init_) {
-      std::cerr << "[difftest] dlsym failed: missing required symbols\n";
-      return false;
-    }
-
-    difftest_init_(0);
-
-    std::vector<uint8_t> pmem(kPmemSize, 0);
-    size_t max_words = kPmemSize / sizeof(uint32_t);
-    size_t word_cnt = pmem_words.size() < max_words ? pmem_words.size() : max_words;
-    size_t copy_bytes = word_cnt * sizeof(uint32_t);
-    if (copy_bytes > 0) {
-      std::memcpy(pmem.data(), pmem_words.data(), copy_bytes);
-    }
-    difftest_memcpy_(kPmemBase, pmem.data(), pmem.size(), kToRef);
-
-    DifftestCPUState boot = {};
-    boot.pc = entry_pc;
-    boot.csr.mstatus = 0x1800u;
-    boot.csr.mtvec = 0x0u;
-    boot.csr.mepc = 0x0u;
-    boot.csr.mcause = 0x0u;
-    difftest_regcpy_(&boot, kToRef);
-    last_ref_state_ = boot;
-    has_last_ref_state_ = true;
-
-    enabled_ = true;
-    std::cout << "[difftest] enabled, image bytes copied=" << pmem.size()
-              << "\n";
-    return true;
-  }
-
-  bool enabled() const { return enabled_; }
-
-  bool step_and_check(uint64_t cycle, uint32_t pc, uint32_t inst,
-                      const std::array<uint32_t, 32> &rf_before,
-                      const std::array<uint32_t, 32> &rf_after) {
-    if (!enabled_) return true;
-
-    DifftestCPUState ref_before = {};
-    difftest_regcpy_(&ref_before, kToDut);
-    if (ref_before.pc != pc) {
-      std::cerr << "[difftest] pc mismatch before exec at cycle " << cycle
-                << " commit_pc=0x" << std::hex << pc << " ref_pc=0x"
-                << ref_before.pc << std::dec << "\n";
-      return false;
-    }
-
-    difftest_exec_(1);
-
-    DifftestCPUState ref_after = {};
-    difftest_regcpy_(&ref_after, kToDut);
-    last_ref_state_ = ref_after;
-    has_last_ref_state_ = true;
-
-    uint32_t mmio_load_rd = 0;
-    bool ignore_mmio_load_rd = decode_mmio_load_rd(inst, rf_before, mmio_load_rd);
-
-    for (int reg = 0; reg < 16; reg++) {
-      if (ignore_mmio_load_rd && reg == static_cast<int>(mmio_load_rd)) continue;
-      if (ref_after.gpr[reg] != rf_after[reg]) {
-        std::cerr << "[difftest] x" << reg << " mismatch at cycle " << cycle
-                  << " pc=0x" << std::hex << pc << " inst=0x" << inst
-                  << ": dut=0x" << rf_after[reg] << " ref=0x"
-                  << ref_after.gpr[reg] << std::dec << "\n";
-        return false;
-      }
-    }
-
-    if (ignore_mmio_load_rd && mmio_load_rd != 0) {
-      ref_after.gpr[mmio_load_rd] = rf_after[mmio_load_rd];
-      difftest_regcpy_(&ref_after, kToRef);
-      last_ref_state_ = ref_after;
-    }
-
-    return true;
-  }
-
-  bool check_arch_state(uint64_t cycle, const std::array<uint32_t, 32> &rf_after,
-                        const DUTCSRState &dut_csr) {
-    if (!enabled_ || !has_last_ref_state_) return true;
-
-    for (int reg = 0; reg < 16; reg++) {
-      if (last_ref_state_.gpr[reg] != rf_after[reg]) {
-        std::cerr << "[difftest] x" << reg
-                  << " mismatch at cycle-end " << cycle
-                  << ": dut=0x" << std::hex << rf_after[reg]
-                  << " ref=0x" << last_ref_state_.gpr[reg] << std::dec << "\n";
-        return false;
-      }
-    }
-
-    if (last_ref_state_.csr.mtvec != dut_csr.mtvec) {
-      std::cerr << "[difftest] mtvec mismatch at cycle-end " << cycle
-                << ": dut=0x" << std::hex << dut_csr.mtvec << " ref=0x"
-                << last_ref_state_.csr.mtvec << std::dec << "\n";
-      return false;
-    }
-    if (last_ref_state_.csr.mepc != dut_csr.mepc) {
-      std::cerr << "[difftest] mepc mismatch at cycle-end " << cycle
-                << ": dut=0x" << std::hex << dut_csr.mepc << " ref=0x"
-                << last_ref_state_.csr.mepc << std::dec << "\n";
-      return false;
-    }
-    if (last_ref_state_.csr.mstatus != dut_csr.mstatus) {
-      std::cerr << "[difftest] mstatus mismatch at cycle-end " << cycle
-                << ": dut=0x" << std::hex << dut_csr.mstatus << " ref=0x"
-                << last_ref_state_.csr.mstatus << std::dec << "\n";
-      return false;
-    }
-    if (last_ref_state_.csr.mcause != dut_csr.mcause) {
-      std::cerr << "[difftest] mcause mismatch at cycle-end " << cycle
-                << ": dut=0x" << std::hex << dut_csr.mcause << " ref=0x"
-                << last_ref_state_.csr.mcause << std::dec << "\n";
-      return false;
-    }
-
-    return true;
-  }
-
-  ~Difftest() {
-    handle_ = nullptr;
-  }
-
- private:
-  using difftest_memcpy_t = void (*)(uint32_t, void *, size_t, bool);
-  using difftest_regcpy_t = void (*)(void *, bool);
-  using difftest_exec_t = void (*)(uint64_t);
-  using difftest_init_t = void (*)(int);
-
-  static constexpr bool kToDut = false;
-  static constexpr bool kToRef = true;
-
-  static int32_t sext12(uint32_t imm12) {
-    return static_cast<int32_t>(imm12 << 20) >> 20;
-  }
-
-  static bool is_mmio_addr(uint32_t addr) {
-    return addr >= kMmioBase && addr <= kMmioEnd;
-  }
-
-  static bool decode_mmio_load_rd(uint32_t inst,
-                                  const std::array<uint32_t, 32> &rf_before,
-                                  uint32_t &rd_out) {
-    uint32_t opcode = inst & 0x7fu;
-    if (opcode != 0x03u) return false;  // LOAD
-
-    uint32_t rd = (inst >> 7) & 0x1fu;
-    uint32_t rs1 = (inst >> 15) & 0x1fu;
-    uint32_t imm12 = (inst >> 20) & 0xfffu;
-    int32_t imm = sext12(imm12);
-
-    if (rs1 >= rf_before.size()) return false;
-    uint32_t addr = rf_before[rs1] + static_cast<uint32_t>(imm);
-    if (!is_mmio_addr(addr)) return false;
-    if (rd == 0 || rd >= 16) return false;
-
-    rd_out = rd;
-    return true;
-  }
-
-  void *handle_ = nullptr;
-  difftest_memcpy_t difftest_memcpy_ = nullptr;
-  difftest_regcpy_t difftest_regcpy_ = nullptr;
-  difftest_exec_t difftest_exec_ = nullptr;
-  difftest_init_t difftest_init_ = nullptr;
-
-  DifftestCPUState last_ref_state_ = {};
-  bool has_last_ref_state_ = false;
-  bool enabled_ = false;
-};
 
 struct UnifiedMem {
   std::vector<uint32_t> pmem_words;
@@ -649,7 +314,7 @@ static void reset(Vtb_triathlon *top, MemSystem &mem, VerilatedVcdC *tfp,
 
 int main(int argc, char **argv) {
   Verilated::commandArgs(argc, argv);
-  SimArgs args = parse_args(argc, argv);
+  npc::SimArgs args = npc::parse_args(argc, argv);
 
   if (args.img_path.empty()) {
     std::cerr << "Usage: " << argv[0]
@@ -681,7 +346,7 @@ int main(int argc, char **argv) {
   }
 #endif
 
-  Difftest difftest;
+  npc::Difftest difftest;
   if (!args.difftest_so.empty()) {
     if (!difftest.init(args.difftest_so, mem.mem.pmem_words, kPmemBase)) {
       if (tfp) tfp->close();
@@ -1666,7 +1331,7 @@ int main(int argc, char **argv) {
         pending_flush_penalty = false;
       }
       if (difftest.enabled()) {
-        DUTCSRState dut_csr = {};
+        npc::DUTCSRState dut_csr = {};
         dut_csr.mtvec = top->dbg_csr_mtvec_o;
         dut_csr.mepc = top->dbg_csr_mepc_o;
         dut_csr.mstatus = top->dbg_csr_mstatus_o;
